break apart _lr
