+++ tag/gym/envs/_lr/env.py
@@ -16,7 +16,7 @@ import torch
 # from .legged_robot_config import LeggedRobotCfg
 
 
-class LeggedRobot(BaseTask):
+class LeggedRobot:
     def __init__(self, cfg: LeggedRobotCfg, sim_device, headless):
         """Parses the provided config file,
             calls create_sim() (which creates, simulation, terrain and environments),
@@ -123,30 +123,6 @@ class LeggedRobot(BaseTask):
         if self.debug_viz:
             self._draw_debug_vis()
 
-    def check_base_pos_out_of_bound(self):
-        """Check if the base position is out of the terrain bounds"""
-        x_out_of_bound = (self.base_pos[:, 0] >= self.terrain_x_range[1]) | (
-            self.base_pos[:, 0] <= self.terrain_x_range[0]
-        )
-        y_out_of_bound = (self.base_pos[:, 1] >= self.terrain_y_range[1]) | (
-            self.base_pos[:, 1] <= self.terrain_y_range[0]
-        )
-        out_of_bound_buf = x_out_of_bound | y_out_of_bound
-        envs_idx = out_of_bound_buf.nonzero(as_tuple=False).flatten()
-        # reset base position to initial position
-        self.base_pos[envs_idx] = self.base_init_pos
-        self.base_pos[envs_idx] += self.env_origins[envs_idx]
-        self.robot.set_pos(self.base_pos[envs_idx], zero_velocity=False, envs_idx=envs_idx)
-
-    def check_termination(self):
-        """Check if environments need to be reset"""
-        self.reset_buf = torch.any(
-            torch.norm(self.link_contact_forces[:, self.termination_indices, :], dim=-1) > 1.0,
-            dim=1,
-        )
-        self.time_out_buf = self.episode_length_buf > self.max_episode_length  # no terminal reward for time-outs
-        self.reset_buf |= self.time_out_buf
-
     def reset_idx(self, env_ids):
         """Reset some environments.
             Calls self._reset_dofs(env_ids), self._reset_root_states(env_ids), and self._resample_commands(env_ids)
@@ -202,70 +178,6 @@ class LeggedRobot(BaseTask):
         # send timeout info to the algorithm
         if self.cfg.env.send_timeouts:
             self.extras["time_outs"] = self.time_out_buf
-
-    def compute_reward(self):
-        """Compute rewards
-        Calls each reward function which had a non-zero scale (processed in self._prepare_reward_function())
-        adds each terms to the episode sums and to the total reward
-        """
-        self.rew_buf[:] = 0.0
-        for i in range(len(self.reward_functions)):
-            name = self.reward_names[i]
-            rew = self.reward_functions[i]() * self.reward_scales[name]
-            self.rew_buf += rew
-            self.episode_sums[name] += rew
-        if self.cfg.rewards.only_positive_rewards:
-            self.rew_buf[:] = torch.clip(self.rew_buf[:], min=0.0)
-        # add termination reward after clipping
-        if "termination" in self.reward_scales:
-            rew = self._reward_termination() * self.reward_scales["termination"]
-            self.rew_buf += rew
-            self.episode_sums["termination"] += rew
-
-    def compute_observations(self):
-        """Computes observations"""
-        self.obs_buf = torch.cat(
-            (
-                self.base_lin_vel * self.obs_scales.lin_vel,  # 3
-                self.base_ang_vel * self.obs_scales.ang_vel,  # 3
-                self.projected_gravity,  # 3
-                self.commands[:, :3] * self.commands_scale,  # 3
-                (self.dof_pos - self.default_dof_pos) * self.obs_scales.dof_pos,  # num_dofs
-                self.dof_vel * self.obs_scales.dof_vel,  # num_dofs
-                self.actions,  # num_actions
-            ),
-            dim=-1,
-        )
-        # add perceptive inputs if not blind
-        if self.cfg.terrain.measure_heights:
-            heights = (
-                torch.clip(
-                    self.base_pos[:, 2].unsqueeze(1) - 0.5 - self.measured_heights,
-                    -1,
-                    1.0,
-                )
-                * self.obs_scales.height_measurements
-            )
-            self.obs_buf = torch.cat((self.obs_buf, heights), dim=-1)
-
-        # add noise if needed
-        if self.add_noise:
-            self.obs_buf += (2 * torch.rand_like(self.obs_buf) - 1) * self.noise_scale_vec
-
-        if self.num_privileged_obs is not None:
-            self.privileged_obs_buf = torch.cat(
-                (
-                    self.base_lin_vel * self.obs_scales.lin_vel,
-                    self.base_ang_vel * self.obs_scales.ang_vel,
-                    self.projected_gravity,
-                    self.commands[:, :3] * self.commands_scale,
-                    (self.dof_pos - self.default_dof_pos) * self.obs_scales.dof_pos,
-                    self.dof_vel * self.obs_scales.dof_vel,
-                    self.actions,
-                    self.last_actions,
-                ),
-                dim=-1,
-            )
 
     def create_sim(self):
         """Creates simulation, terrain and evironments"""
@@ -380,22 +292,6 @@ class LeggedRobot(BaseTask):
         # set small commands to zero
         self.commands[env_ids, :2] *= (torch.norm(self.commands[env_ids, :2], dim=1) > 0.2).unsqueeze(1)
 
-    def _compute_torques(self, actions):
-        # control_type = 'P'
-        actions_scaled = actions * self.cfg.control.action_scale
-        torques = (
-            self.batched_p_gains * (actions_scaled + self.default_dof_pos - self.dof_pos)
-            - self.batched_d_gains * self.dof_vel
-        )
-        return torques
-
-    def _compute_target_dof_pos(self, actions):
-        # control_type = 'P'
-        actions_scaled = actions * self.cfg.control.action_scale
-        target_dof_pos = actions_scaled + self.default_dof_pos
-
-        return target_dof_pos
-
     def _reset_dofs(self, envs_idx):
         """Resets DOF position and velocities of selected environmments
         Positions are randomly selected within 0.5:1.5 x default positions.
@@ -457,65 +353,6 @@ class LeggedRobot(BaseTask):
         self.base_ang_vel[envs_idx] = gs_rand_float(-0.5, 0.5, (len(envs_idx), 3), self.device)
         base_vel = torch.concat([self.base_lin_vel[envs_idx], self.base_ang_vel[envs_idx]], dim=1)
         self.robot.set_dofs_velocity(velocity=base_vel, dofs_idx_local=[0, 1, 2, 3, 4, 5], envs_idx=envs_idx)
-
-    def _push_robots(self):
-        """Random pushes the robots. Emulates an impulse by setting a randomized base velocity."""
-        if self.push_interval_s > 0 and not self.debug:
-            max_push_vel_xy = self.cfg.domain_rand.max_push_vel_xy
-            # in Genesis, base link also has DOF, it's 6DOF if not fixed.
-            dofs_vel = self.robot.get_dofs_velocity()  # (num_envs, num_dof) [0:3] ~ base_link_vel
-            push_vel = gs_rand_float(-max_push_vel_xy, max_push_vel_xy, (self.num_envs, 2), self.device)
-            push_vel[((self.common_step_counter + self.env_identities) % int(self.push_interval_s / self.dt) != 0)] = 0
-            dofs_vel[:, :2] += push_vel
-            self.robot.set_dofs_velocity(dofs_vel)
-
-    def _update_terrain_curriculum(self, env_ids):
-        """Implements the game-inspired curriculum.
-
-        Args:
-            env_ids (List[int]): ids of environments being reset
-        """
-        # Implement Terrain curriculum
-        if not self.init_done:
-            # don't change on initial reset
-            return
-        distance = torch.norm(self.base_pos[env_ids, :2] - self.env_origins[env_ids, :2], dim=1)
-        # robots that walked far enough progress to harder terains
-        move_up = distance > self.utils_terrain.env_length / 2
-        # robots that walked less than half of their required distance go to simpler terrains
-        move_down = (
-            distance < torch.norm(self.commands[env_ids, :2], dim=1) * self.max_episode_length_s * 0.5
-        ) * ~move_up
-        self.terrain_levels[env_ids] += 1 * move_up - 1 * move_down
-        # Robots that solve the last level are sent to a random one
-        self.terrain_levels[env_ids] = torch.where(
-            self.terrain_levels[env_ids] >= self.max_terrain_level,
-            torch.randint_like(self.terrain_levels[env_ids], self.max_terrain_level),
-            torch.clip(self.terrain_levels[env_ids], 0),
-        )  # (the minumum level is zero)
-        self.env_origins[env_ids] = self.terrain_origins[self.terrain_levels[env_ids], self.terrain_types[env_ids]]
-
-    def update_command_curriculum(self, env_ids):
-        """Implements a curriculum of increasing commands
-
-        Args:
-            env_ids (List[int]): ids of environments being reset
-        """
-        # If the tracking reward is above 80% of the maximum, increase the range of commands
-        if (
-            torch.mean(self.episode_sums["tracking_lin_vel"][env_ids]) / self.max_episode_length
-            > 0.8 * self.reward_scales["tracking_lin_vel"]
-        ):
-            self.command_ranges["lin_vel_x"][0] = np.clip(
-                self.command_ranges["lin_vel_x"][0] - 0.5,
-                -self.cfg.commands.max_curriculum,
-                0.0,
-            )
-            self.command_ranges["lin_vel_x"][1] = np.clip(
-                self.command_ranges["lin_vel_x"][1] + 0.5,
-                0.0,
-                self.cfg.commands.max_curriculum,
-            )
 
     def _get_noise_scale_vec(self, cfg):
         """Sets a vector used to scale the noise added to the observations.
@@ -629,54 +466,6 @@ class LeggedRobot(BaseTask):
         self.robot.set_dofs_kp(self.p_gains, self.motor_dofs)
         self.robot.set_dofs_kv(self.d_gains, self.motor_dofs)
 
-    def _prepare_reward_function(self):
-        """Prepares a list of reward functions, whcih will be called to compute the total reward.
-        Looks for self._reward_<REWARD_NAME>, where <REWARD_NAME> are names of all non zero reward scales in the cfg.
-        """
-        # remove zero scales + multiply non-zero ones by dt
-        for key in list(self.reward_scales.keys()):
-            scale = self.reward_scales[key]
-            if scale == 0:
-                self.reward_scales.pop(key)
-            else:
-                self.reward_scales[key] *= self.dt
-        # prepare list of functions
-        self.reward_functions = []
-        self.reward_names = []
-        for name, scale in self.reward_scales.items():
-            if name == "termination":
-                continue
-            self.reward_names.append(name)
-            name = "_reward_" + name
-            self.reward_functions.append(getattr(self, name))
-
-        # reward episode sums
-        self.episode_sums = {
-            name: torch.zeros(
-                self.num_envs,
-                dtype=gs.tc_float,
-                device=self.device,
-                requires_grad=False,
-            )
-            for name in self.reward_scales.keys()
-        }
-
-    def _create_heightfield(self):
-        """Adds a heightfield terrain to the simulation, sets parameters based on the cfg."""
-        self.terrain = self.scene.add_entity(
-            gs.morphs.Terrain(
-                pos=(-self.cfg.terrain.border_size, -self.cfg.terrain.border_size, 0.0),
-                horizontal_scale=self.cfg.terrain.horizontal_scale,
-                vertical_scale=self.cfg.terrain.vertical_scale,
-                height_field=self.utils_terrain.height_field_raw,
-            )
-        )
-        self.height_samples = (
-            torch.tensor(self.utils_terrain.heightsamples)
-            .view(self.utils_terrain.tot_rows, self.utils_terrain.tot_cols)
-            .to(self.device)
-        )
-
     def _create_envs(self):
         """Creates environments:
         1. loads the robot URDF/MJCF asset, create entity
@@ -750,48 +539,6 @@ class LeggedRobot(BaseTask):
         if self.cfg.domain_rand.randomize_com_displacement:
             self._randomize_com_displacement(np.arange(self.num_envs))
 
-    def _randomize_friction(self, env_ids=None):
-        """Randomize friction of all links"""
-        min_friction, max_friction = self.cfg.domain_rand.friction_range
-
-        solver = self.rigid_solver
-
-        ratios = (
-            gs.rand((len(env_ids), 1), dtype=float).repeat(1, solver.n_geoms) * (max_friction - min_friction)
-            + min_friction
-        )
-        solver.set_geoms_friction_ratio(ratios, torch.arange(0, solver.n_geoms), env_ids)
-
-    def _randomize_base_mass(self, env_ids=None):
-        """Randomize base mass"""
-        min_mass, max_mass = self.cfg.domain_rand.added_mass_range
-        base_link_id = 1
-        added_mass = gs.rand((len(env_ids), 1), dtype=float) * (max_mass - min_mass) + min_mass
-        self.rigid_solver.set_links_mass_shift(
-            added_mass,
-            [
-                base_link_id,
-            ],
-            env_ids,
-        )
-
-    def _randomize_com_displacement(self, env_ids):
-        min_displacement, max_displacement = self.cfg.domain_rand.com_displacement_range
-        base_link_id = 1
-
-        com_displacement = (
-            gs.rand((len(env_ids), 1, 3), dtype=float) * (max_displacement - min_displacement) + min_displacement
-        )
-        # com_displacement[:, :, 0] -= 0.02
-
-        self.rigid_solver.set_links_COM_shift(
-            com_displacement,
-            [
-                base_link_id,
-            ],
-            env_ids,
-        )
-
     def _parse_cfg(self, cfg):
         self.dt = self.cfg.control.dt
         if self.cfg.sim.use_implicit_controller:  # use embedded PD controller
@@ -830,210 +577,3 @@ class LeggedRobot(BaseTask):
         self.scene.draw_debug_spheres(
             height_points[0, :], radius=0.03, color=(0, 0, 1, 0.7)
         )  # only draw for the first env
-
-    def _get_env_origins(self):
-        """Sets environment origins. On rough terrain the origins are defined by the terrain platforms.
-        Otherwise create a grid.
-        """
-        if self.cfg.terrain.mesh_type in ["heightfield"]:
-            self.custom_origins = True
-            self.env_origins = torch.zeros(self.num_envs, 3, device=self.device, requires_grad=False)
-            # put robots at the origins defined by the terrain
-            max_init_level = self.cfg.terrain.max_init_terrain_level
-            if not self.cfg.terrain.curriculum:
-                max_init_level = self.cfg.terrain.num_rows - 1
-            self.terrain_levels = torch.randint(0, max_init_level + 1, (self.num_envs,), device=self.device)
-            self.terrain_types = torch.div(
-                torch.arange(self.num_envs, device=self.device),
-                (self.num_envs / self.cfg.terrain.num_cols),
-                rounding_mode="floor",
-            ).to(torch.long)
-            self.max_terrain_level = self.cfg.terrain.num_rows
-            self.terrain_origins = torch.from_numpy(self.utils_terrain.env_origins).to(self.device).to(torch.float)
-            self.env_origins[:] = self.terrain_origins[self.terrain_levels, self.terrain_types]
-        else:
-            self.custom_origins = False
-            self.env_origins = torch.zeros(self.num_envs, 3, device=self.device, requires_grad=False)
-            # create a grid of robots
-            num_cols = np.floor(np.sqrt(self.num_envs))
-            num_rows = np.ceil(self.num_envs / num_cols)
-            xx, yy = torch.meshgrid(torch.arange(num_rows), torch.arange(num_cols))
-            # plane has limited size, we need to specify spacing base on num_envs, to make sure all robots are within the plane
-            # restrict envs to a square of [plane_length/2, plane_length/2]
-            spacing = min(
-                (self.cfg.terrain.plane_length / 2) / (num_rows - 1),
-                (self.cfg.terrain.plane_length / 2) / (num_cols - 1),
-            )
-            self.env_origins[:, 0] = spacing * xx.flatten()[: self.num_envs]
-            self.env_origins[:, 1] = spacing * yy.flatten()[: self.num_envs]
-            self.env_origins[:, 2] = 0.0
-            self.env_origins[:, 0] -= self.cfg.terrain.plane_length / 4
-            self.env_origins[:, 1] -= self.cfg.terrain.plane_length / 4
-
-    def _init_height_points(self):
-        """Returns points at which the height measurments are sampled (in base frame)
-
-        Returns:
-            [torch.Tensor]: Tensor of shape (num_envs, self.num_height_points, 3)
-        """
-        y = torch.tensor(self.cfg.terrain.measured_points_y, device=self.device, requires_grad=False)
-        x = torch.tensor(self.cfg.terrain.measured_points_x, device=self.device, requires_grad=False)
-        grid_x, grid_y = torch.meshgrid(x, y)
-
-        self.num_height_points = grid_x.numel()
-        points = torch.zeros(
-            self.num_envs,
-            self.num_height_points,
-            3,
-            device=self.device,
-            requires_grad=False,
-        )
-        points[:, :, 0] = grid_x.flatten()
-        points[:, :, 1] = grid_y.flatten()
-        return points
-
-    def _get_heights(self, env_ids=None):
-        """Samples heights of the terrain at required points around each robot.
-            The points are offset by the base's position and rotated by the base's yaw
-
-        Args:
-            env_ids (List[int], optional): Subset of environments for which to return the heights. Defaults to None.
-
-        Raises:
-            NameError: [description]
-
-        Returns:
-            [type]: [description]
-        """
-        if self.cfg.terrain.mesh_type == "plane":
-            return torch.zeros(
-                self.num_envs,
-                self.num_height_points,
-                device=self.device,
-                requires_grad=False,
-            )
-        elif self.cfg.terrain.mesh_type == "none":
-            raise NameError("Can't measure height with terrain mesh type 'none'")
-
-        if env_ids:
-            points = quat_apply_yaw(
-                self.base_quat[env_ids].repeat(1, self.num_height_points),
-                self.height_points[env_ids],
-            ) + (self.base_pos[env_ids, :3]).unsqueeze(1)
-        else:
-            points = quat_apply_yaw(self.base_quat.repeat(1, self.num_height_points), self.height_points) + (
-                self.base_pos[:, :3]
-            ).unsqueeze(1)
-
-        points += self.cfg.terrain.border_size
-        points = (points / self.cfg.terrain.horizontal_scale).long()
-        px = points[:, :, 0].view(-1)
-        py = points[:, :, 1].view(-1)
-        px = torch.clip(px, 0, self.height_samples.shape[0] - 2)
-        py = torch.clip(py, 0, self.height_samples.shape[1] - 2)
-
-        heights1 = self.height_samples[px, py]
-        heights2 = self.height_samples[px + 1, py]
-        heights3 = self.height_samples[px, py + 1]
-        heights = torch.min(heights1, heights2)
-        heights = torch.min(heights, heights3)
-
-        return heights.view(self.num_envs, -1) * self.cfg.terrain.vertical_scale
-
-    # ------------ reward functions----------------
-    def _reward_lin_vel_z(self):
-        # Penalize z axis base linear velocity
-        return torch.square(self.base_lin_vel[:, 2])
-
-    def _reward_ang_vel_xy(self):
-        # Penalize xy axes base angular velocity
-        return torch.sum(torch.square(self.base_ang_vel[:, :2]), dim=1)
-
-    def _reward_orientation(self):
-        # Penalize non flat base orientation
-        return torch.sum(torch.square(self.projected_gravity[:, :2]), dim=1)
-
-    def _reward_base_height(self):
-        # Penalize base height away from target
-        base_height = torch.mean(self.base_pos[:, 2].unsqueeze(1) - self.measured_heights, dim=1)
-        rew = torch.square(base_height - self.cfg.rewards.base_height_target)
-        return rew
-
-    def _reward_torques(self):
-        # Penalize torques
-        return torch.sum(torch.square(self.torques), dim=1)
-
-    def _reward_dof_vel(self):
-        # Penalize dof velocities
-        return torch.sum(torch.square(self.dof_vel), dim=1)
-
-    def _reward_dof_acc(self):
-        # Penalize dof accelerations
-        return torch.sum(torch.square((self.last_dof_vel - self.dof_vel) / self.dt), dim=1)
-
-    def _reward_action_rate(self):
-        # Penalize changes in actions
-        return torch.sum(torch.square(self.last_actions - self.actions), dim=1)
-
-    def _reward_collision(self):
-        # Penalize collisions on selected bodies
-        return torch.sum(
-            1.0 * (torch.norm(self.link_contact_forces[:, self.penalized_indices, :], dim=-1) > 0.1),
-            dim=1,
-        )
-
-    def _reward_termination(self):
-        # Terminal reward / penalty
-        return self.reset_buf * ~self.time_out_buf
-
-    def _reward_dof_pos_limits(self):
-        # Penalize dof positions too close to the limit
-        out_of_limits = -(self.dof_pos - self.dof_pos_limits[:, 0]).clip(max=0.0)  # lower limit
-        out_of_limits += (self.dof_pos - self.dof_pos_limits[:, 1]).clip(min=0.0)
-        return torch.sum(out_of_limits, dim=1)
-
-    # def _reward_dof_vel_limits(self):
-    #     # Penalize dof velocities too close to the limit
-    #     # clip to max error = 1 rad/s per joint to avoid huge penalties
-    #     return torch.sum((torch.abs(self.dof_vel) - self.dof_vel_limits*self.cfg.rewards.soft_dof_vel_limit).clip(min=0., max=1.), dim=1)
-
-    def _reward_torque_limits(self):
-        # penalize torques too close to the limit
-        return torch.sum(
-            (torch.abs(self.torques) - self.torque_limits * self.cfg.rewards.soft_torque_limit).clip(min=0.0),
-            dim=1,
-        )
-
-    def _reward_tracking_lin_vel(self):
-        # Tracking of linear velocity commands (xy axes)
-        lin_vel_error = torch.sum(torch.square(self.commands[:, :2] - self.base_lin_vel[:, :2]), dim=1)
-        return torch.exp(-lin_vel_error / self.cfg.rewards.tracking_sigma)
-
-    def _reward_tracking_ang_vel(self):
-        # Tracking of angular velocity commands (yaw)
-        ang_vel_error = torch.square(self.commands[:, 2] - self.base_ang_vel[:, 2])
-        return torch.exp(-ang_vel_error / self.cfg.rewards.tracking_sigma)
-
-    def _reward_feet_air_time(self):
-        # Reward long steps
-        contact = self.link_contact_forces[:, self.feet_indices, 2] > 1.0
-        contact_filt = torch.logical_or(contact, self.last_contacts)
-        self.last_contacts = contact
-        first_contact = (self.feet_air_time > 0.0) * contact_filt
-        self.feet_air_time += self.dt
-        rew_airTime = torch.sum(
-            (self.feet_air_time - 0.5) * first_contact, dim=1
-        )  # reward only on first contact with the ground
-        rew_airTime *= torch.norm(self.commands[:, :2], dim=1) > 0.1  # no reward for zero command
-        self.feet_air_time *= ~contact_filt
-        return rew_airTime
-
-    def _reward_stand_still(self):
-        # Penalize motion at zero commands
-        return torch.sum(torch.abs(self.dof_pos - self.default_dof_pos), dim=1) * (
-            torch.norm(self.commands[:, :2], dim=1) < 0.1
-        )
-
-    def _reward_dof_close_to_default(self):
-        # Penalize dof position deviation from default
-        return torch.sum(torch.square(self.dof_pos - self.default_dof_pos), dim=1)
